% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_mat.R
\name{alluvial_mat}
\alias{alluvial_mat}
\title{alluvial_mat}
\usage{
alluvial_mat(new.projections, ct_anno)
}
\arguments{
\item{new.projections}{new projection to visualize}

\item{ct_anno}{Cell type annotation}
}
\value{
A dataframe that can be used to plot an alluvial
}
\description{
Function to compute alluvial matrix
}
\examples{
pr <- projectR(data=p.ESepiGen4c1l$mRNA.Seq,Patterns=AP.RNAseq6l3c3t$Amean, 
AnnotationObj=map.ESepiGen4c1l,IDcol="GeneSymbols",full = TRUE)
alluvial_mat(pr,ct_anno = c(rep('hESC_Ectoderm',2),rep('hESC_Endoderm',3),
rep('hESC_Mesoderm',2),rep('hESC_Cntrl',2)))
}
