% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_mat.R
\name{auc_mat}
\alias{auc_mat}
\title{auc_mat}
\usage{
auc_mat(labels = NA, weights = NA)
}
\arguments{
\item{labels}{a vector of labels whose length is equal to the number of rows in the weight matrix}

\item{weights}{a matrix of weights from projection analysis}
}
\value{
A matrix of AUC values for each set of weights classifying each label.
}
\description{
Calculates AUC values for each set of weights for each label and outputs the results as a matrix
}
