% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clust.R
\name{bindex}
\alias{bindex}
\title{A function to creat new report with required files
copied and created to work on}
\usage{
bindex(dataset, na.rm = FALSE, verbose = TRUE)
}
\arguments{
\item{dataset}{A matrix or data.frame, usually with columns representing samples
and rows representing genes or proteins.}

\item{na.rm}{Logical. If remove na before calculating bi. If False, bi will not be calculated for
any row  with NA,}

\item{verbose}{A logical value; should the function output a stream of information while it is working?}
}
\value{
Returns a data frame containing six columns, with the rows corresponding to
 the rows of the original data set.
 The columns contain the four parameters from the normal mixture model (mu1, mu2, sigma, and pi)
 along with the standardized distance delta and the bimodal index BI.
}
\description{
Calculate Bimodal Index
}
\note{
update from bimodalIndex {BimodalIndex}
}
\references{
Wang J, Wen S, Symmans WF, Pusztai L, Coombes KR.
The bimodality index: A criterion for discovering and ranking bimodal signatures from cancer gene expression profiling data.
Cancer Informatics, 2009 Aug 5; 7:199–216.
}

