% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{cluster2pattern-class}
\alias{cluster2pattern-class}
\title{cluster2pattern}
\description{
class of cluster2pattern output.
}
\section{Slots}{

\describe{
\item{\code{clusterMatrix}}{matrix of continous values for projection that is output of cluster2pattern function}
}}

