% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\docType{methods}
\name{cluster2pattern,kmeans-method}
\alias{cluster2pattern,kmeans-method}
\title{cluster2pattern (kmeans)}
\usage{
<<<<<<< HEAD:man/cluster2pattern-kmeans-method.Rd
\S4method{cluster2pattern}{kmeans}(clusters, NP = NA, Data = NA)
=======
cluster2pattern.kmeans(clusters = NA, NP = NA, Data = NA)
>>>>>>> 68e7dc7b88caed3ca6bf64468c75fe11aef8577f:man/cluster2pattern.kmeans.Rd
}
\arguments{
\item{clusters}{an kmeans cluster object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\value{
An object of class 'pclust' containing pattern weights corresponding for each cluster.
}
\description{
Function to make patterns of continuous weights from kmeans clusters.
}
\examples{
 k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
 cluster2pattern(clusters=k.RNAseq6l3c3t,NP=22,Data=p.RNAseq6l3c3t)

}
