% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\name{cluster2pattern.hclust}
\alias{cluster2pattern.hclust}
\title{cluster2pattern function (hclust)}
\usage{
cluster2pattern.hclust(clusters, NP = NA, Data = NA)
}
\arguments{
\item{clusters}{an hclust object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\value{
An object of class 'pclust' containing pattern weights corresponding for each cluster.
}
\description{
Function to make patterns of continuous weights from clusters.
}
\examples{
 distp <- dist(p.RNAseq6l3c3t)
 hc.RNAseq6l3c3t <- hclust(distp)
 cluster2pattern(clusters=hc.RNAseq6l3c3t,NP=22,Data=p.RNAseq6l3c3t)
}
