% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\name{cluster2pattern.kmeans}
\alias{cluster2pattern.kmeans}
\title{cluster2pattern (kmeans)}
\usage{
\method{cluster2pattern}{kmeans}(clusters = NA, NP = NA, Data = NA)
}
\arguments{
\item{clusters}{an kmeans cluster object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\description{
Function to make patterns of continuous weights from kmeans clusters.
}
\examples{
\dontrun{
 cluster2pattern(clusters,NP,Data)
}
}
