% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\name{cluster2pattern}
\alias{cluster2pattern}
\title{cluster2pattern}
\usage{
cluster2pattern(clusters = NA, NP = NA, Data = NA)
}
\arguments{
\item{clusters}{an cluster object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\description{
Function to make patterns of continuous weights from clusters.
}
\examples{
\dontrun{
 cluster2pattern(clusters,NP,Data)
}
}

