% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlotR.R
\name{clusterPlotR.kmeans}
\alias{clusterPlotR.kmeans}
\title{clusterPlotR (kmeans)}
\usage{
\method{clusterPlotR}{kmeans}(cData = NA, cls = NA, x = NA, NC = NA,
  annoIndx = NA, label = NULL, ...)
}
\arguments{
\item{cData}{data used to get clusters}

\item{cls}{a kmeans object}

\item{x}{a vector of length equal to number of samples to use for plotting}

\item{NC}{vector of integers indicating which clusters to use}

\item{annoIndx}{vector indxing into subsets for plotting}

\item{...}{additional parameters for plotting. ex. pch,cex,col,labels, xlab, etc.}

\item{labels}{character vector to use for plotting text, defaults is NULL}
}
\description{
plotting function for kmeans clusters
}
\examples{
\dontrun{
 clusterPlotR(cData=p, cls=pk, x=jitter(pd$days), col=pd$colors)
}
}
