% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{clusterPlotR}
\alias{clusterPlotR}
\title{Generic clusterPlotR function}
\usage{
clusterPlotR(cData, cls, x, NC, ...)
}
\arguments{
\item{cData}{data used to get clusters}

\item{cls}{a cluster (kmeans or hclust) object}

\item{x}{a vector of length equal to number of samples to use for plotting}

\item{NC}{vector of integers indicating which clusters to use}
}
\value{
A plot of the mean behavior for each cluster
}
\description{
Generic clusterPlotR function
}
\examples{
\dontrun{
 k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
 clusterPlotR(p.RNAseq6l3c3t, cls=k.RNAseq6l3c3t, NC=1,x=pd.RNAseq6l3c3t$days, col=pd.RNAseq6l3c3t$color)
}
}
