% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_me_up.R
\name{confirm_directory}
\alias{confirm_directory}
\title{Prompt to confirm current working directory}
\usage{
confirm_directory(projectname)
}
\arguments{
\item{projectname}{A character string of the project title}
}
\value{
A logical whether or not to set up the in the current working directory
}
\description{
This is a helper function, in order to confirm the current working directory
}
