% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{correlateR-class}
\alias{correlateR-class}
\title{correlateR}
\description{
class of correlateR output.
}
\section{Slots}{

\describe{
\item{\code{corM}}{correlation matrix obtained from correlateR}
}}

