% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{create a new project}
\usage{
create_project(name, dir = name, path = getwd(), google_proj = name,
  google_path = "projects", repo = name, server = "gitlab.com",
  user = getOption("gitlab_user"))
}
\arguments{
\item{name}{project name}

\item{dir}{project directory}

\item{path}{local project location}

\item{google_proj}{google drive location}

\item{google_path}{google drive projects path

 
before this function can be called, there must be a directory called 
\code{local} in the working directory containing two authorization files
and a configuration directory:

* \code{gitlab_token} a personal access token from gitlab with full api
  scope
* \code{drive_cache} a binary cache file authorizing \code{googledrive} to 
  access your drive.  this can be created by running:
  \code{googledrive::drive_auth(cache = "local/drive_cache")}
* \code{.gd} a configuration directory for Emmanuel
  Odeke's excellent \code{drive} software (https://github.com/odeke-em/drive)
  This can be created by running \code{drive init} in a shell in the 
  \code{local} directory.  Do not run \code{drive} again in this directory.}

\item{repo}{gitlab repository name}

\item{url}{gitlab server url}
}
\description{
create a new project
}
