% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_projectR}
\alias{create_projectR}
\title{Create a new R project}
\usage{
create_projectR(
  path,
  git = FALSE,
  repo = NULL,
  languages = c("R"),
  renv = FALSE
)
}
\arguments{
\item{path}{The path to create the project in.}

\item{git}{Logical value indicating whether to initialize a Git repository.}

\item{repo}{The name of a GitHub repository to clone to create the project.}

\item{languages}{A character vector of languages to include in the gitignore file.}

\item{renv}{Logical value indicating whether to initialize an R environment.}
}
\value{
A message indicating whether the project was created successfully.
}
\description{
This function creates a new R project with a default directory structure,
gitignore templates, and a setup.R script in the R directory.
}
