% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{data_file_path}
\alias{data_file_path}
\title{Data Directory File Path}
\usage{
data_file_path(path)
}
\arguments{
\item{path}{relative path from data directory to file}
}
\value{
absolute path to file
}
\description{
data_file_path returns an absolute path to a file in the
data directory set by \code{\link{set_data_directory}}. The data
directory can be different for each collaborator of the project.
}
