% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload_dir.R
\name{drive_upload_dir}
\alias{drive_upload_dir}
\title{upload a local directory to google drive, recursively}
\usage{
drive_upload_dir(from, to = from, ignore = NULL, regex = FALSE,
  is_new = FALSE)
}
\arguments{
\item{from}{local directory path}

\item{to}{remote directory path (not parent)

if \code{to} exists, the contents of \code{from} will be copied into 
\code{to}, otherwise \code{to} will first be created.  if a file in 
\code{from}
already exists in \code{to}, it will be updated, otherwise it will be 
uploaded.  subdirectories will be recursed}
}
\description{
upload a local directory to google drive, recursively
}
