% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{external_file_path}
\alias{external_file_path}
\title{Return File From External Project Directory}
\usage{
external_file_path(path)
}
\arguments{
\item{path}{relative path from external directory to file}
}
\value{
absolute path to file
}
\description{
external_file_path returns an absolute path to a file in the
external directory set by \code{\link{set_external_directory}}. The external
directory can be different for each collaborator of the project.
}
