% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneMatchR.R
\docType{methods}
\name{geneMatchR}
\alias{geneMatchR}
\alias{geneMatchR,ANY-method}
\title{Generic geneMatchR function}
\usage{
geneMatchR(data1, AnnotationObj, IDcol = "GeneSymbol", data2,
  merge = FALSE, ...)

\S4method{geneMatchR}{ANY}(data1 = NA, AnnotationObj = NA,
  IDcol = "GeneSymbol", data2 = NA, merge = FALSE)
}
\arguments{
\item{data1}{a dataset of genes by samples}

\item{AnnotationObj}{an annotation object for data1. If NA, the rownames of data will be used.}

\item{IDcol}{the column of AnnotationData object corresponding to identifiers matching the rownames of data2}

\item{data2}{dataset with rownames to be matched}

\item{merge}{logical indicating wether or not to merged data sets}

\item{...}{Additional arguments to geneMatchR}
}
\value{
A list of genes (intersection) in both datasets. (if merge = TRUE, Also returns merged data.)
}
\description{
Matches genes accross datasets
}
\examples{
geneMatchR(data1=p.RNAseq6l3c3t,AnnotationObj=map.ESepiGen4c1l,
IDcol="GeneSymbols",data2=p.ESepiGen4c1l$mRNA.Seq)
}
