% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneMatchR.R
\name{geneMatchR}
\alias{geneMatchR}
\title{geneMatchR}
\usage{
geneMatchR(data1 = NA, AnnotionObj = NA, IDcol = "GeneSymbol",
  data2 = NA, merge = FALSE)
}
\arguments{
\item{data1}{a dataset of genes by samples}

\item{AnnotionObj}{an annotion object for data1. If NA, the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the rownames of data2}

\item{data2}{dataset with rownames to be matched}

\item{merge}{logical indicating wether or not to merged data sets}
}
\description{
Matches genes accross datasets.
}

