% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getTSNE}
\alias{getTSNE}
\title{getTSNE}
\usage{
getTSNE(projection, axis = 2, ...)
}
\arguments{
\item{projection}{martrix, a projection generated from projectR}

\item{axis}{integer, either 1 umap of projection or 2 for umap of transpose of projection}

\item{...}{addtional arguments passed to tsne}
}
\description{
Function to provide tSNE of projection
}
\examples{
projection <- projectR(data=p.ESepiGen4c1l$mRNA.Seq,loadings=AP.RNAseq6l3c3t$Amean,
dataNames = map.ESepiGen4c1l[["GeneSymbols"]], full = TRUE)
projectionTSNE <- getTSNE(projection)
}
