% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_me_wings.R
\name{give_me_wings}
\alias{give_me_wings}
\title{Create folder structure}
\usage{
give_me_wings()
}
\value{
A set of folders created in project root
}
\description{
This function will create a folder structure starting from the project root
that is necessary for a data science project
}
