% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{initialize,cluster2pattern-method}
\alias{initialize,cluster2pattern-method}
\title{Constructor for cluster2pattern}
\usage{
\S4method{initialize}{cluster2pattern}(.Object, clusterMatrix, ...)
}
\arguments{
\item{.Object}{clusterMatrix object}

\item{clusterMatrix}{matrix of continous values for projection that is output of cluster2pattern function}

\item{...}{additional arguments to intialize cluster2pattern}
}
\value{
initialized cluster2pattern object
}
\description{
Constructor for cluster2pattern
}
