% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{initialize,correlateR-method}
\alias{initialize,correlateR-method}
\title{Constructor for correlateR}
\usage{
\S4method{initialize}{correlateR}(.Object, corM, ...)
}
\arguments{
\item{.Object}{correlateR object}

\item{corM}{correlation matrix obtained from correlateR}

\item{...}{additional arguments to intialize correlateR}
}
\value{
initialized correlateR object
}
\description{
Constructor for correlateR
}
