% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{initialize,pclust-method}
\alias{initialize,pclust-method}
\title{Constructor for pclust}
\usage{
\S4method{initialize}{pclust}(.Object, patterns, ...)
}
\arguments{
\item{.Object}{pclust object}

\item{patterns}{patterns found from clusters (either hclust or kmeans object)) using cluster2pattern}

\item{...}{Additional arguments to initalize pclust}
}
\value{
initialized plclust object
}
\description{
Constructor for pclust
}
