% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{initialize,pclustHclust-method}
\alias{initialize,pclustHclust-method}
\title{Constructor for pclust}
\usage{
\S4method{initialize}{pclustHclust}(.Object, hclust, ...)
}
\arguments{
\item{.Object}{pclust object}

\item{hclust}{hclust object passed to cluster2pattern function}

\item{...}{additional arguments to intialize pclust}
}
\value{
initialized plclustHclust object
}
\description{
Constructor for pclust
}
