% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{initialize,pclustKmeans-method}
\alias{initialize,pclustKmeans-method}
\title{Constructor for pclustKmeans}
\usage{
\S4method{initialize}{pclustKmeans}(.Object, kmeans, ...)
}
\arguments{
\item{.Object}{pclustKmeans object}

\item{kmeans}{kmeans object passed to \code{cluster2pattern} generic}

\item{...}{Additional arguments to \code{cluster2pattern} generic}
}
\value{
initialized plclustKmeans object
}
\description{
Constructor for pclustKmeans
}
