% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{initialize,rotatoR-method}
\alias{initialize,rotatoR-method}
\title{Constructor for rotatoR}
\usage{
\S4method{initialize}{rotatoR}(.Object, rotatedM, ...)
}
\arguments{
\item{.Object}{rotatoR object}

\item{rotatedM}{rotated matrix from rotatoR function}

\item{...}{additional arguments to intialize rotatoR}
}
\value{
initialized rotatoR object
}
\description{
Constructor for rotatoR
}
