% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectoR.R
\name{intersectoR.default}
\alias{intersectoR.default}
\title{intersectoR (Base)}
\usage{
intersectoR.default(pSet1 = NA, pSet2 = NA, pval = 0.05,
  full = FALSE, k = NULL)
}
\arguments{
\item{pSet1}{a list for a set of patterns where each entry is a set of genes associated with a single pattern}

\item{pSet2}{a list for a second set of patterns where each entry is a set of genes associated with a single pattern}

\item{pval}{the maximum p-value considered significant}

\item{full}{logical indicating whether to return full data frame of signigicantly overlapping sets. Default is false will return summary matrix.}

\item{k}{numeric giving cut height for hclust objects, if vector arguments will be applied to pSet1 and pSet2 in that order}

\item{pSet1}{a list for a set of patterns where each entry is a set of genes associated with a single pattern}

\item{pSet2}{a list for a second set of patterns where each entry is a set of genes associated with a single pattern}

\item{pval}{the maximum p-value considered significant}

\item{full}{logical indicating whether to return full data frame of signigicantly overlapping sets. Default is false will return summary matrix.}

\item{k}{cut height for hclust objects}
}
\value{
A list containing: Overlap matrix, overlap index, and overlapping sets.

A list containing: Overlap matrix, overlap index, and overlapping sets.
}
\description{
a function to find and test the intersecting values of two sets of lists, presumably the genes associated with patterns in two different datasets.

a function to find and test the intersecting values of two sets of lists, presumably the genes associated with patterns in two different datasets.
}
\examples{
ESepiGen4c1lmRNASeq <- p.ESepiGen4c1l$mRNA.Seq
rownames(ESepiGen4c1lmRNASeq) <- map.ESepiGen4c1l$GeneSymbols
k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
k.ESepiGen4c1l<-kmeans(ESepiGen4c1lmRNASeq,10)
intersectoR(k.RNAseq6l3c3t, k.ESepiGen4c1l, pval=.05)
\dontrun{
 intersectoR(pSet1, pSet2, pval=.05)
}
}
