% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectoR.R
\name{intersectoR.kmeans}
\alias{intersectoR.kmeans}
\title{intersectoR (Kmeans)}
\usage{
intersectoR.kmeans(pSet1 = NA, pSet2 = NA, pval = 0.05,
  full = FALSE)
}
\arguments{
\item{pSet1}{a kmeans object}

\item{pSet2}{a second kmeans object}

\item{pval}{the maximum p-value considered significant}

\item{full}{logical indicating whether to return full data frame of signigicantly overlapping sets. Default is false will return summary matrix.}

\item{k}{cut height for hclust objects, not used with kmeans}
}
\value{
A list containing: Overlap matrix, overlap index, and overlapping sets.
}
\description{
a function to find and test the intersecting values of two sets of kmeans clusters, presumably the genes associated with clusters in two different datasets.
}
\examples{
ESepiGen4c1lmRNASeq <- p.ESepiGen4c1l$mRNA.Seq
rownames(ESepiGen4c1lmRNASeq) <- map.ESepiGen4c1l$GeneSymbols
k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
k.ESepiGen4c1l<-kmeans(ESepiGen4c1lmRNASeq,10)
intersectoR(k.RNAseq6l3c3t, k.ESepiGen4c1l, pval=.05)
}
