% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{intersectoR}
\alias{intersectoR}
\title{Generic intersectoR function}
\usage{
intersectoR(pSet1, pSet2, pval, ...)
}
\arguments{
\item{pSet1}{a list for a set of patterns where each entry is a set of genes associated with a single pattern}

\item{pSet2}{a list for a second set of patterns where each entry is a set of genes associated with a single pattern}

\item{pval}{the maximum p-value considered significant}

\item{k}{numeric giving cut height for hclust objects, if vector arguments will be applied to pSet1 and pSet2 in that order}
}
\value{
A list containing: Overlap matrix, overlap index, and overlapping sets.
}
\description{
a function to find and test the intersecting values of two sets of lists, presumably the genes associated with patterns in two different datasets.
}
\examples{
 k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
 k.ESepiGen4c1l<-kmeans(p.ESepiGen4c1l$mRNA.Seq,10)
intersectoR(k.RNAseq6l3c3t, k.ESepiGen4c1l, pval=.05)

}
