% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{KL}
\alias{KL}
\title{KL Divergence between Two Probability Distance Matrix}
\usage{
KL(dist_a, dist_b)
}
\arguments{
\item{a}{The first embeded probability distance matrix}

\item{b}{The second embeded probability distance matrix}
}
\value{
KL Divergence between two probability distance matrix
}
\description{
Compute the KL Divergence between two embedded probability distance matrix
}
\examples{

dist_a = probDistance(matrix(rnorm(100 * 3), nrow = 100) )
dist_b = probDistance(matrix(rnorm(100 * 5), nrow = 100) )
out = KL(dist_a, dist_b)

}
\keyword{Difference}
\keyword{Divergence,}
\keyword{KL}
