% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littlegit.R
\name{littlegit}
\alias{littlegit}
\title{Lightweight Git project bootstrap}
\usage{
littlegit(path, git = TRUE, repo = NULL)
}
\arguments{
\item{path}{Character. Path to project root}

\item{git}{Logical. Whether to use Git}

\item{repo}{Character or NULL. GitHub repo spec "user/repo"}
}
\value{
Invisibly returns normalized project path
}
\description{
Initializes a project directory with optional Git support,
optionally cloning from GitHub.
}
