% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littlegitignore.R
\name{littlegitignore}
\alias{littlegitignore}
\title{Create .gitignore with optional templates}
\usage{
littlegitignore(path, languages = c("R"), renv = FALSE)
}
\description{
Create .gitignore with optional templates
}
