% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{map.RNAseq6l3c3t}
\alias{map.RNAseq6l3c3t}
\title{RNAseqing from human 3 iPS & 3 ES cell lines
in 3 experimental condition at 3 time points}
\format{
A data frames with 108 rows and 54 variables:
}
\usage{
data(map.RNAseq6l3c3t)
}
\description{
map.RNAseq6l3c3 contains gene annotations for polyA bulk sequencing of 6 cell lines
in 3 experimental condition at 3 time points.
}
\keyword{datasets}
