% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matprod_par.R
\name{matprod.par}
\alias{matprod.par}
\title{Parallel calculation}
\usage{
matprod.par(cl, A, B)
}
\arguments{
\item{cl}{integer, number of cluster}

\item{A}{Matrix}

\item{B}{Matrix}
}
\description{
Matrix calculation parallel execution
}
\details{
The idea is that we will distribute the rows of a matrix A for example
on the cores and we will make the matrix product of each group of
rows with another matrix B
}
\examples{
\dontrun{
ncores=8
cl <- makeCluster(ncores)
A <- matrix(round(rnorm(16^2),1),16)
B <- t(A)+4
matprod.par(cl,A,B)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
