% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{microglial_counts}
\alias{microglial_counts}
\title{log-normalized count data from microglial cells in the p6 mouse cortex.}
\format{
A gene (rows) by cell (column) matrix
}
\usage{
data(microglial_counts)
}
\description{
log-normalized count data from microglial cells in the p6 mouse cortex.
}
\keyword{datasets}
