% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{multivariateAnalysisR_seurat_test}
\alias{multivariateAnalysisR_seurat_test}
\title{Truncated Seurat Object with latent space projection
done to unspecified cells in different stages for multivariateAnalysisR analysis}
\format{
A Seurat Object with 31034 observations of 4 variables in meta.data:
}
\usage{
data(multivariateAnalysisR_seurat_test)
}
\description{
Truncated Seurat Object with latent space projection
done to unspecified cells in different stages for multivariateAnalysisR analysis
}
\keyword{datasets}
