% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{new_report}
\alias{new_report}
\title{A function to creat new report with required files
copied and created to work on}
\usage{
new_report(report_name, report_path = getwd(), echo = TRUE)
}
\arguments{
\item{report_name}{the name of report}

\item{report_path}{path of the report. default: getwd()}

\item{echo, }{logic. Show verbose information}
}
\value{
NA
}
\description{
Create a new report
}

