% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{p.ESepiGen4c1l}
\alias{p.ESepiGen4c1l}
\title{RNAseqing and ChIPSeq of matched genes in differentiated human iPS cells}
\format{
p.ESepiGen4c1l is a list of 6 data frames each with with 93 rows and between 4 and 9 variables:
}
\usage{
data(p.ESepiGen4c1l)
}
\description{
p.ESepiGen4c1l contains log2(RPKM + 1) values for polyA bulk sequencing
and log2 counts of normalized ChIPSeq reads of 1 cell lines with 2
replicates in 4 experimental conditions at a single time point.
}
\references{
1. Gifford, C. A. et al. Transcriptional and epigenetic dynamics
during specification of human embryonic stem cells. Cell 153, 1149-1163 (2013).
}
\keyword{datasets}
