% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{pclustHclust-class}
\alias{pclustHclust-class}
\title{pclustHclust}
\description{
defines class of cluster2pattern with hclust input
}
\section{Slots}{

\describe{
\item{\code{hclust}}{hclust input to cluster2pattern}
}}

