% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{pclustKmeans-class}
\alias{pclustKmeans-class}
\title{pclustKmeans}
\description{
defines class of cluster2pattern output with kmeans input
}
\section{Slots}{

\describe{
\item{\code{kmeans}}{kmeans object used as input to cluster2pattern}
}}

