% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pd.RNAseq6l3c3t}
\alias{pd.RNAseq6l3c3t}
\title{RNAseqing from human 3 iPS & 3 ES cell lines
in 3 experimental condition at 3 time points}
\format{
A data frames with 54 rows and 38 variables:
}
\usage{
data(pd.RNAseq6l3c3t)
}
\description{
pd.RNAseq6l3c3t contains sample phenotype and experimental information
for polyA bulk sequencing of 6 cell lines in 3 experimental condition at 3 time points.
}
\keyword{datasets}
