% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotConfidenceIntervals}
\alias{plotConfidenceIntervals}
\title{plotConfidenceIntervals}
\usage{
plotConfidenceIntervals(confidence_intervals, interval_name = c("low", "high"))
}
\arguments{
\item{confidence_intervals}{a vector with names of loading rows. Defaults to rownames.}

\item{interval_name}{names of columns that contain the low and high estimates, respectively}
}
\description{
Generate point and line confidence intervals from provided estimates
}
