% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plotVolcano}
\usage{
plotVolcano(stats, metadata, FC, pvalue, title)
}
\arguments{
\item{stats}{data frame with differential expression statistics}

\item{metadata}{#metadata from pdVolcano}

\item{FC}{Fold change threshold}

\item{pvalue}{p value threshold}

\item{title}{plot title}
}
\description{
Volcano plotting function
}
