% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{probDistance}
\alias{probDistance}
\title{Probability Distance Matrix}
\usage{
probDistance(x, sigma_est = NULL)
}
\arguments{
\item{x}{A matrix of dimension n * p, which is the embedded feature or a selection of genes}

\item{sigma_est}{Take value of NULL or a manually set number}
}
\value{
A computed probability distance matrix
}
\description{
Compute the probability distance matrix
}
\examples{
x = matrix(rnorm(100 * 3), nrow = 100) 
dist_mat = probDistance(x)

}
\keyword{distance}
\keyword{matrix}
\keyword{probability}
