% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_start.R
\name{proj_start}
\alias{proj_start}
\title{Start a new project}
\usage{
proj_start(proj_dir, data_dir = NULL)
}
\arguments{
\item{proj_dir}{character. The directory where the new project
should exist}

\item{data_dir}{character. The directory where the raw data exist. If
\code{NULL}, the data directory is created as a subdirectory.}
}
\description{
\code{proj_start} creates a directory with the essential files for
a new project. This is basically a stripped-down version of
\code{workflowr::wflow_start} with some choices that are tailored
to my own preferences.
}
\details{
\code{proj_start} populates the chosen directory with the
following files:

\preformatted{|--- .gitignore
|--- <directory>.Rproj
|--- analysis/
|--- data/
|--- drafts/
|--- results/
|--- source/
|--- README.md
}

\code{analysis/} is intended to contain all the source R Markdown
files that implement the analyses for the project.

\code{data/} will either be a symbolic link to an external data
directory, or a subdirectory.

\code{drafts/} will contain all references and relevant paper drafts

\code{results/} will contain results exported by the analysis files

\code{source/} will contain bare scripts (typically containing
functions sourced by the full analysis files)

By default \code{.gitignore} indicates to Git to ignore files usually
ignored in an R Project (\code{.Rhistory}, \code{.Rdata}, ect), as
well as the data directory
}
