% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{create a new project and return its \code{project} object}
\usage{
project(name, path = getwd(), dir = name, repo = name,
  server = "gitlab.com",
  gitlab_token = read_token("local/gitlab_token"), drive_proj = name,
  drive_path = "projects", drive_cache = "local/drive_cache",
  user = getOption("gitlab_user"))
}
\arguments{
\item{name}{project name}

\item{path}{local project location (parent directory)}

\item{dir}{project directory (\code{name} by convention)}

\item{repo}{gitlab repository name (\code{name} by convention)}

\item{server}{gitlab server url}

\item{gitlab_token}{gitlab personal access token}

\item{drive_proj}{google drive location}

\item{drive_path}{google drive projects path}

\item{drive_cache}{google drive cache location

 
before this function can be called, there must be a directory called 
\code{local} in the working directory containing two authorization files

* \code{gitlab_token} a personal access token from gitlab with full api
  scope
* \code{drive_cache} a binary cache file authorizing \code{googledrive} to 
  access your drive.  this can be created by running:
  \code{googledrive::drive_auth(cache = "local/drive_cache")}}
}
\description{
create a new project and return its \code{project} object
}
