% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectoR.R
\name{projectoR.CoGAPS}
\alias{projectoR.CoGAPS}
\title{Projection function (CoGAPS)}
\usage{
\method{projectoR}{CoGAPS}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, full = FALSE)
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{a CoGAPS object}

\item{NP}{vector of integers indicating which columns of Patterns object to use. The default of NP = NA will use entire matrix.}

\item{full}{logical indicating whether to return the full model solution. By default only the new pattern object is returned.}
}
\description{
for use with object of class CoGAPS
}
\examples{
\dontrun{
   projectoR(data=D,Patterns=AP,PatternData=D)
}
}

