% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectoR.R
\name{projectoR.hclust}
\alias{projectoR.hclust}
\title{<Projection function (hierachical clustering)>}
\usage{
\method{projectoR}{hclust}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, PatternData = NA,
  full = FALSE, ...)
}
\arguments{
\item{data}{a dataset to be projected onto}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{an hclust object}

\item{NP}{number of desired patterns}

\item{PatternData}{data used to make hclust object}

\item{full}{logical indicating whether to return the full model solution. By default only the new pattern object is returned.}

\item{...}{additional parameters for hclust output}
}
\description{
<for use with object of class hclust>
}
\examples{
\dontrun{
   projectoR(data=D,Patterns=cls,PatternData=D)
}
}

