% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectoR.R
\name{projectoR.pclust}
\alias{projectoR.pclust}
\title{Projection function (clustering)}
\usage{
\method{projectoR}{pclust}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, full = FALSE)
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{an Pclust object from the cluster2pattern function}

\item{NP}{number of desired patterns}

\item{full}{logical indicating whether to return the full model solution. By default only the new pattern object is returned.}
}
\description{
for use with object of class Pclust
}
\examples{
\dontrun{
   projectoR(data=D,Patterns=cls,PatternData=D)
}
}
