% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectR.R
\name{projectR.correlateR}
\alias{projectR.correlateR}
\title{Projection function (correlateR)}
\usage{
\method{projectR}{correlateR}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, full = FALSE,
  model = NA)
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{an correlateR object}

\item{NP}{can be used to select for "NegativeCOR" or "PositiveCOR" list from correlateR class obj containing both. By default is NA}

\item{full}{logical indicating whether to return the full clustering information.  By default only the new pattern object is returned.}

\item{model}{# optional arguements to choose method for projection}
}
\value{
A matrix of sample weights for each input pattern. (if full=TRUE, full model solution is returned)
}
\description{
for use with object of class corR
}
\examples{
 c.RNAseq6l3c3t<-correlateR(genes="T", dat=p.RNAseq6l3c3t, threshtype="N", threshold=10, absR=TRUE)
 cor.ESepiGen4c1l<-projectR(data=p.ESepiGen4c1l$mRNA.Seq,Patterns=c.RNAseq6l3c3t,NP="PositiveCOR",
                                   AnnotionObj=map.ESepiGen4c1l,IDcol="GeneSymbols")

}
