% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectoR.R
\name{projectR.kmeans}
\alias{projectR.kmeans}
\title{<Projection function (kmeans clustering)>}
\usage{
\method{projectR}{kmeans}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, PatternData = NA,
  full = FALSE, ...)
}
\arguments{
\item{data}{a dataset to be projected onto}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{a kmeans object}

\item{NP}{vector of integers indicating which columns of Patterns object to use. The default of NP = NA will use entire matrix.}

\item{PatternData}{data used to make kmeans object}

\item{full}{logical indicating whether to return the full model solution. By default only the new pattern object is returned.}

\item{...}{additional inputs to}
}
\description{
<for use with object of class kmeans>
}
\examples{
\dontrun{
   projectR(data=D,Patterns=cls,PatternData=D)
}
}
