% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectR.R
\name{projectR.list}
\alias{projectR.list}
\title{Projection function (list for CoGAPS)}
\usage{
projectR.list(data = NA, AnnotationObj = NA, IDcol = "GeneSymbol",
  Patterns = NA, NP = NA, full = FALSE, model = NA)
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotationObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{an CoGAPS object}

\item{NP}{can be used to select for "NegativeCOR" or "PositiveCOR" list from correlateR class obj containing both. By default is NA}

\item{full}{logical indicating whether to return the full clustering information.  By default only the new pattern object is returned.}

\item{model}{# optional arguements to choose method for projection}
}
\value{
A matrix of sample weights for each input pattern. (if full=TRUE, full model solution is returned)
}
\description{
for use with object of class corR
}
\examples{
 projectR(data=p.ESepiGen4c1l$mRNA.Seq,Patterns=AP.RNAseq6l3c3t,
               AnnotationObj=map.ESepiGen4c1l,IDcol="GeneSymbols")

}
