% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectR.R
\name{projectR.prcomp}
\alias{projectR.prcomp}
\title{Projection function (PCA)}
\usage{
\method{projectR}{prcomp}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, full = FALSE)
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{an prcomp object with a rotation matrix of genes by PCs}

\item{NP}{range of PCs to project. The default of NP = NA will use entire matrix.}

\item{full}{logical indicating whether to return the percent variance accounted for by each projected PC. By default only the new pattern object is returned.}
}
\description{
for use with object of class prcomp
}
