% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{projectR}
\alias{projectR}
\title{Generic projectR function}
\usage{
projectR(data, Patterns, ...)
}
\arguments{
\item{data}{Target dataset into which you will prject}

\item{Patterns}{Patterns learned from source dataset}

\item{...}{Additional arguments to projectR}
}
\value{
A matrix of sample weights for each input pattern. (if full=TRUE, full model solution is returned)
}
\description{
Generic projectR function
}
\examples{
        projectR(data=p.RNAseq6l3c3t,Patterns=AP.RNAseq6l3c3t)

}
