% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectR.R
\name{projectR.rotatoR}
\alias{projectR.rotatoR}
\title{Projection function (rotatoR objects)}
\usage{
<<<<<<< HEAD
<<<<<<< HEAD
\method{projectR}{rotatoR}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, full = FALSE,
  model = NA)
=======
projectR.rotatoR(data = NA, AnnotionObj = NA, IDcol = "GeneSymbol",
  Patterns = NA, NP = NA, full = FALSE, model = NA)
>>>>>>> 68e7dc7b88caed3ca6bf64468c75fe11aef8577f
=======
projectR.rotatoR(data = NA, AnnotionObj = NA, IDcol = "GeneSymbol",
  Patterns = NA, NP = NA, full = FALSE, model = NA)
>>>>>>> master
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{an rotatoR object with a rotation matrix of genes by new PCs}

\item{NP}{range of PCs to project. The default of NP = NA will use entire matrix.}

\item{full}{logical indicating whether to return the percent variance accounted for by each projected PC. By default only the new pattern object is returned.}

\item{model}{# optional arguements to choose method for projection}
}
\value{
A matrix of sample weights for each input pattern. (if full=TRUE, full model solution is returned)
}
\description{
for use with object of class rotatoR
}
\examples{
# 
#   Do not run, this will break the build.  Fix the generics instead.
#  
# pca.RNAseq6l3c3t<-prcomp(t(p.RNAseq6l3c3t))
# r.RNAseq6l3c3t<-rotatoR(1,1,-1,-1,pca.RNAseq6l3c3t$x[,1:2])
#  pca.ESepiGen4c1l<-projectR(data=p.ESepiGen4c1l$mRNA.Seq,Patterns=r.RNAseq6l3c3t,
#                         AnnotionObj=map.ESepiGen4c1l,IDcol="GeneSymbols")

}
