% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{render}
\alias{here}
\alias{map}
\alias{walk}
\alias{make_clean_names}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{here}{\code{\link[here]{here}}}

  \item{janitor}{\code{\link[janitor]{make_clean_names}}}

  \item{purrr}{\code{\link[purrr]{map}}, \code{\link[purrr:map]{walk}}}

  \item{rmarkdown}{\code{\link[rmarkdown]{render}}}
}}

