% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{keep trying to evaluate an expression until it succeeds}
\usage{
retry(expr, sleep = 1, max_tries = 10)
}
\arguments{
\item{expr}{expression to be evaluated}

\item{sleep}{seconds to sleep between tries}

\item{max_tries}{maximum number of tries before we throw up our hands in
disgust}
}
\value{
the value of the successfully evaluated expression.
}
\description{
keep trying to evaluate an expression until it succeeds
}
\details{
some functions, especially those calling curl, may fail 
non-deterministically.  this function repeatedly wraps \code{expr} in a 
new promise and evaluates it until it succeeds or \code{max_tries} 
is exceeded
}
