% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq6l3c3.R
\name{RNAseq6l3c3t}
\alias{RNAseq6l3c3t}
\title{RNAseqing of developmental genes in human ES/iPS cell differentiation}
\format{p.RNAseq6l3c3 is data frame with 108 rows and 54 variables:}
\description{
p.RNAseq6l3c3 contains log2(RPKM + 1) values for polyA bulk sequencing of 6 cell lines
in 3 experimental condition at 3 time points.
}
\details{
map.RNAseq6l3c3 contains gene annotations

pd.RNAseq6l3c3 contains sample phenotype and experimental information
}
