% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_stat.R
\name{row-test}
\alias{row-test}
\alias{row_ftest}
\alias{row_ttest}
\title{A wrapper to conduct t-test/F-test row-by-row}
\usage{
row_ttest(data, class, pair = NULL, na.rm = TRUE, var.equal = FALSE, base)

row_ftest(data, class, na.rm = TRUE, var.equal = FALSE, base)
}
\arguments{
\item{data}{matrix/data.frame}

\item{class}{vector/factor for class label}

\item{pair}{paired vector}

\item{na.rm}{logical FALSE. should remove NA before test?}

\item{var.equal}{Logical FALSE}

\item{base}{the base for input data to calculate fold change}
}
\value{
data.frame with statistics, df and pvalues,fc
}

