% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{set_data_directory}
\alias{set_data_directory}
\title{Set Data Directory}
\usage{
set_data_directory(path)
}
\arguments{
\item{path}{absolute path to data directory}
}
\description{
set_data_directory sets a directory external to the project
directory with the \code{\link[config]} package. The absolute path to the data
directory is stored in a config.yml file that is added to the .gitignore. The
data directory can be different for each collaborator of the project.
}
