% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{set_external_directory}
\alias{set_external_directory}
\title{Set External Project Directory}
\usage{
set_external_directory(path)
}
\arguments{
\item{path}{absolute path to external directory}
}
\description{
set_external_directory sets a directory external to the project directory
with the config package. The absolute path to the external directory is
stored in a config.yml file that is added to the .gitignore file if it
exists. The external directory can be different for each collaborator of the
project.
}
