% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_me_up.R
\name{show_structure}
\alias{show_structure}
\title{Output composed project structure description}
\usage{
show_structure(projectname)
}
\arguments{
\item{projectname}{A character string of the project title}
}
\value{
A console output of the project structure
}
\description{
This is a helper function, to output a composed structure using the
directory tree with the corresponding description and the projectname, that
was specified by the user.
}
