% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
calculate the sigmoid function
}
\details{
It represents the x value in the distribution function of the logistic law.
}
\examples{
\dontrun{
sigmoid(15)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
