% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{similarityMeasure}
\alias{similarityMeasure}
\title{Measurement of Similarity}
\usage{
similarityMeasure(latent1, latent2, sigma = 1)
}
\arguments{
\item{latent1}{First embedded latent space}

\item{latent2}{Second embedded latent space}

\item{sigma}{The sigma used in the maximum mean discrepancy (MMD) between two latent spaces}
}
\value{
Computed MMD between two latent spaces
}
\description{
Compute the similarity of two embedded latent spaces
}
\examples{

latent1 = matrix(rnorm(100 * 3), nrow = 100)
latent2 = matrix(rnorm(500 * 3), nrow = 500)
out = similarityMeasure(latent1, latent2, 1)

}
\keyword{MMD}
\keyword{Similarity,}
