% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_01_comparison.R
\name{t-test}
\alias{row_t}
\alias{row_t_paired}
\alias{row_t_unequal}
\alias{t-test}
\title{conduct t-test row-by-row}
\usage{
row_t(m, v, na.rm = FALSE)

row_t_unequal(m, v, na.rm = FALSE)

row_t_paired(m, v, pf, na.rm = FALSE)
}
\arguments{
\item{m}{matrix}

\item{v}{vector for class label}

\item{na.rm}{logical FALSE}

\item{pf}{paired vector}
}
\value{
list with statistics, df and pvalues
}
\examples{
x = matrix(rnorm(120),20,6)
y = rep(1:2,each=3)
z = rep(1:3,2)
t1 = row_t(x,y)
t2 = row_t_unequal(x,y)
t3 = row_t_paired(x,y,z)
t3
}

