% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_mat.R
\name{alluvial_mat}
\alias{alluvial_mat}
\title{alluvial_mat}
\usage{
alluvial_mat(new.projections, ct_anno)
}
\arguments{
\item{new.projections}{new projection to visualize}

\item{ct_anno}{Cell type annotation}
}
\value{
A matrix
}
\description{
Function to compute alluvial matrix
}
