% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvialMat.R
\name{alluvialMat}
\alias{alluvialMat}
\title{alluvialMat}
\usage{
alluvialMat(projection, annotations, annotationName = "Cell type",
  annotationType = "Cell", plot = TRUE, minPropExplained = 0.75)
}
\arguments{
\item{projection}{a projection generated from projectR, ensure that full = TRUE while generating projection}

\item{annotationName}{a character for naming the annotations collectively, default is "Cell type"}

\item{annotationType}{a character indicating the type of data annotated, default is "Cell"}

\item{plot}{logical indicating whether to return the alluvial plot, default is TRUE}

\item{minPropExplained}{}

\item{annotation}{a character vector of annotations for the data}
}
\value{
A matrix to generate alluvial plots
}
\description{
Function to provide alluvial matrix for generating alluvial plot
}
\details{
If threshtype is "R" than threshold must be between -1 and 1. Otherwise if top N correlated genes are required, set \code{threshtype}
 as "N" and set \code{threshold} = N, i.e, the number of correlated genes required.
}
\examples{
projection <- projectR(data=p.ESepiGen4c1l$mRNA.Seq,loadings=AP.RNAseq6l3c3t$Amean, 
dataNames = map.ESepiGen4c1l[["GeneSymbols"]], full = TRUE)
alluvialMat(projection,pd.ESepiGen4c1l$Condition)
}
