% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{AP.RNAseq6l3c3t}
\alias{AP.RNAseq6l3c3t}
\title{CoGAPS patterns and genes weights for p.RNAseq6l3c3t}
\format{A list of 12 items}
\usage{
AP.RNAseq6l3c3t
}
\description{
AP.RNAseq6l3c3t contains the output of the gapsRun function in the 
CoGAPS package for data = p.RNAseq6l3c3t
}
\keyword{datasets}
