% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionDriveRfun.R
\name{bonferroniCorrectedDifferences}
\alias{bonferroniCorrectedDifferences}
\title{bonferroniCorrectedDifferences}
\usage{
bonferroniCorrectedDifferences(
  group1,
  group2,
  pvalue,
  diff_weights = NULL,
  mode = "CI"
)
}
\arguments{
\item{group1}{count matrix 1}

\item{group2}{count matrix 2}

\item{pvalue}{significance value to threshold}

\item{diff_weights}{loadings to weight the differential expression}

\item{mode}{statistical approach, confidence intervals(CI) or pvalues(PV)}
}
\description{
Calculate weighted/unweighted mean difference for each gene between 2 groups
}
