% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionDriveRfun.R
\name{bonferroniCorrectedDifferences}
\alias{bonferroniCorrectedDifferences}
\title{bonferroniCorrectedDifferences}
\usage{
bonferroniCorrectedDifferences(
  group1,
  group2,
  diff_weights = NULL,
  mode = "CI",
  pvalue
)
}
\arguments{
\item{group1}{count matrix 1}

\item{group2}{count matrix 2}

\item{diff_weights}{loadings to weight the differential expression between the groups}

\item{mode}{user specified statistical approach, confidence intervals (CI) or pvalues (PV) - default to CI}

\item{pvalue}{significance value to threshold}
}
\description{
Calculate the weighted and unweighted difference in means for each measurement between two groups.
}
