% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionDriveR.R
\name{bonferroniCorrectedDifferences}
\alias{bonferroniCorrectedDifferences}
\title{Calculate the (weighted) difference in means for each measurement between two groups.}
\usage{
bonferroniCorrectedDifferences(group1, group2, diff_weights = NULL, pvalue)
}
\description{
Calculate the (weighted) difference in means for each measurement between two groups.
}
