% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\docType{methods}
\name{cluster2pattern,hclust-method}
\alias{cluster2pattern,hclust-method}
\title{cluster2pattern (hclust)}
\usage{
<<<<<<< HEAD:man/cluster2pattern-hclust-method.Rd
\S4method{cluster2pattern}{hclust}(clusters, NP = NA, Data = NA)
=======
cluster2pattern.hclust(clusters = NA, NP = NA, Data = NA)
>>>>>>> 68e7dc7b88caed3ca6bf64468c75fe11aef8577f:man/cluster2pattern.hclust.Rd
}
\arguments{
\item{clusters}{an hclust object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\value{
An object of class 'pclust' containing pattern weights corresponding for each cluster.
}
\description{
Function to make patterns of continuous weights from hierarchical clusters.
}
\examples{
 h.RNAseq6l3c3t<-hclust(as.dist(1-(cor(t(p.RNAseq6l3c3t),use="pairwise.complete.obs"))))
 cluster2pattern(clusters=h.RNAseq6l3c3t,NP=22,Data=p.RNAseq6l3c3t)

}
