% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\name{cluster2pattern.hclust}
\alias{cluster2pattern.hclust}
\title{cluster2pattern (hclust)}
\usage{
\method{cluster2pattern}{hclust}(clusters = NA, NP = NA, Data = NA)
}
\arguments{
\item{clusters}{an hclust object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\description{
Function to make patterns of continuous weights from hierarchical clusters.
}
\examples{
\dontrun{
 cluster2pattern(clusters,NP,Data)
}
}

