% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2pattern.R
\name{cluster2pattern}
\alias{cluster2pattern}
\title{cluster2pattern}
\usage{
cluster2pattern(clusters = NA, NP = NA, Data = NA)
}
\arguments{
\item{clusters}{an cluster object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}
}
\description{
Function to make patterns of continuous weights from clusters.
}
\examples{
 k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
 cluster2pattern(clusters=k.RNAseq6l3c3t,NP=22,Data=p.RNAseq6l3c3t)  

}
