% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cluster2pattern}
\alias{cluster2pattern}
\title{cluster2pattern}
\usage{
<<<<<<< HEAD
cluster2pattern(clusters, NP, Data)
=======
cluster2pattern(clusters, NP, Data, ...)
>>>>>>> 68e7dc7b88caed3ca6bf64468c75fe11aef8577f
}
\arguments{
\item{clusters}{an cluster object}

\item{NP}{number of desired patterns}

\item{Data}{data used to make clusters object}

\item{...}{Additional arguments to geneMatchR}
}
\value{
An object of class 'pclust' containing pattern weights corresponding for each cluster.
}
\description{
Function to make patterns of continuous weights from clusters.
}
\examples{
 k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
 cluster2pattern(clusters=k.RNAseq6l3c3t,NP=22,Data=p.RNAseq6l3c3t)
}
