% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlotR.R
\name{clusterPlotR}
\alias{clusterPlotR}
\title{clusterPlotR (base)}
\usage{
clusterPlotR(cData = NA, cls = NA, x = NA, NC = NA, annoIndx = NA,
  ...)
}
\arguments{
\item{cData}{data used to get clusters}

\item{cls}{an clustering object}

\item{x}{a vector of length equal to number of samples to use for plotting}

\item{NC}{number of clusters to cut dendrogram into}

\item{annoIndx}{vector indxing into subsets for plotting}

\item{...}{additional parameters for plotting. ex. pch,cex,col,labels, xlab, etc.}
}
\value{
A plot of the mean behavior for each cluster
}
\description{
plotting function for clustering objects
}
\examples{
\dontrun{
 k.RNAseq6l3c3t<-kmeans(p.RNAseq6l3c3t,22)
 clusterPlotR(p.RNAseq6l3c3t, cls=k.RNAseq6l3c3t, NC=2,x=pd.RNAseq6l3c3t$days, col=pd.RNAseq6l3c3t$color)
}
}
