% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{CR.RNAseq6l3c3t}
\alias{CR.RNAseq6l3c3t}
\title{CogapsResult object for p.RNAseq6l3c3t}
\format{
A CogapsResult object
}
\usage{
CR.RNAseq6l3c3t
}
\description{
CR.RNAseq6l3c3t contains the output of the CoGAPS function in the
CoGAPS package for data = p.RNAseq6l3c3t
}
\keyword{datasets}
