% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{cr_microglial}
\alias{cr_microglial}
\title{CogapsResult object for microglial_counts}
\format{
A CogapsResult object
}
\usage{
cr_microglial
}
\description{
cr_microglia contains the output of the CoGAPS function in the
CoGAPS package for data = microglial_counts
}
\keyword{datasets}
