% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveRs.R
\name{driveRs}
\alias{driveRs}
\title{<Projection driveRs function >}
\usage{
driveRs(data = NA, projectedP = NA, design = NA, contrasts = NA,
  NP = NA, full = FALSE, ...)
}
\arguments{
\item{data}{a dataset to be projected onto}

\item{projectedP}{an prcomp object with a rotation matrix of genes by PCs}

\item{design}{design matrix describing experiment}

\item{contrasts}{a matrix of continous values with unique rownames to be projected}

\item{NP}{vector of integers indicating which columns of Patterns object to use. The default of NP = NA will use entire matrix.}

\item{full}{logical indicating whether to return the full model solution. By default only the new pattern object is returned.}

\item{...}{additional arguements to topTable}
}
\description{
<full description>
}
\examples{
\dontrun{
   project(data=D,Patterns=AP)
}
}
\seealso{
topTable, lmFit, eBayes
}
