% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl.R
\name{gitlab_curl}
\alias{gitlab_curl}
\title{do a curl command with default settings for gitlab}
\usage{
gitlab_curl(endpoint, method = "GET", atts = NULL)
}
\description{
do a curl command with default settings for gitlab
}
