% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{glial_counts}
\alias{glial_counts}
\title{log-normalized count data from astrocytes and oligodendrocytes in the p6 mouse cortex.}
\format{
A gene (rows) by cell (column) matrix
}
\usage{
glial_counts
}
\description{
log-normalized count data from astrocytes and oligodendrocytes in the p6 mouse cortex.
}
\keyword{datasets}
