% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{initialize,pclust-method}
\alias{initialize,pclust-method}
\title{Constructor for pclust}
\usage{
\S4method{initialize}{pclust}(.Object, patterns, ...)
}
\arguments{
\item{.Object}{pclust object}

\item{hclust}{hclust object passed to cluster2pattern function}

\item{kmeans}{kmeans object passed to cluster2pattern function}
}
\value{
initialized plclust object
}
\description{
Constructor for pclust
}
