% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{initialize,rotatoR-method}
\alias{initialize,rotatoR-method}
\title{Constructor for rotatoR}
\usage{
\S4method{initialize}{rotatoR}(.Object, clusterMatrix, ...)

\S4method{initialize}{rotatoR}(.Object, clusterMatrix, ...)
}
\arguments{
\item{.Object}{clusterMatrix object}

\item{clusterMatrix}{matrix of continous values for projection that is output of cluster2pattern function}

\item{...}{additional arguments to intialize cluster2pattern}

\item{rotatedM}{rotated matrix from rotatoR function}
}
\value{
initialized rotatoR object

initialized cluster2pattern object
}
\description{
Constructor for rotatoR

Constructor for rotatoR
}
