% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectoR.R
\name{intersectoR.hclust}
\alias{intersectoR.hclust}
\title{intersectoR (hclust)}
\usage{
\method{intersectoR}{hclust}(pSet1 = NA, pSet2 = NA, pval = 0.05,
  full = FALSE, k = NULL)
}
\arguments{
\item{pSet1}{a hclust object}

\item{pSet2}{a second hclust object}

\item{pval}{the maximum p-value considered significant}

\item{full}{logical indicating whether to return full data frame of signigicantly overlapping sets. Default is false will return summary matrix.}

\item{k}{#numeric giving cut height for hclust objects, if vector arguments will be applied to pSet1 and pSet2 in that order}
}
\description{
a function to find and test the intersecting values of two hierarchial clustering objects, presumably the genes associated with clusters in two different datasets.
}
\examples{
\dontrun{
 intersector(pSet1, pSet2, pval=.05, k=c(3,4))
}
}

