% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectoR.R
\name{intersectoR.kmeans}
\alias{intersectoR.kmeans}
\title{intersectoR (Kmeans)}
\usage{
\method{intersectoR}{kmeans}(pSet1 = NA, pSet2 = NA, pval = 0.05,
  full = FALSE, k = NULL)
}
\arguments{
\item{pSet1}{a kmeans object}

\item{pSet2}{a second kmeans object}

\item{pval}{the maximum p-value considered significant}

\item{full}{logical indicating whether to return full data frame of signigicantly overlapping sets. Default is false will return summary matrix.}

\item{k}{cut height for hclust objects, not used with kmeans}
}
\description{
a function to find and test the intersecting values of two sets of kmeans clusters, presumably the genes associated with clusters in two different datasets.
}
\examples{
\dontrun{
 intersectoR(pSet1, pSet2, pval=.05)
}
}

