% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectoR.R
\name{intersectoR}
\alias{intersectoR}
\title{intersectoR (Base)}
\usage{
intersectoR(pSet1 = NA, pSet2 = NA, pval = 0.05, full = FALSE,
  k = NULL)
}
\arguments{
\item{pSet1}{a list for a set of patterns where each entry is a set of genes associated with a single pattern}

\item{pSet2}{a list for a second set of patterns where each entry is a set of genes associated with a single pattern}

\item{pval}{the maximum p-value considered significant}

\item{full}{logical indicating whether to return full data frame of signigicantly overlapping sets. Default is false will return summary matrix.}

\item{k}{numeric giving cut height for hclust objects, if vector arguments will be applied to pSet1 and pSet2 in that order}
}
\description{
a function to find and test the intersecting values of two sets of lists, presumably the genes associated with patterns in two different datasets.
}
\examples{
\dontrun{
 intersectoR(pSet1, pSet2, pval=.05)
}
}
