% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{map.ESepiGen4c1l}
\alias{map.ESepiGen4c1l}
\title{RNAseqing and ChIPSeq of matched genes in differentiated human iPS cells}
\format{
A data frames with 93 rows and 9 variables:
}
\usage{
map.ESepiGen4c1l
}
\description{
map.ESepiGen4c1l contains gene annotations
}
\references{
1. Gifford, C. A. et al. Transcriptional and epigenetic dynamics 
during specification of human embryonic stem cells. Cell 153, 1149-1163 (2013).
}
\keyword{datasets}
