% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.RNAseq6l3c3t.R
\docType{data}
\name{p.RNAseq6l3c3t}
\alias{p.RNAseq6l3c3t}
\title{RNAseqing from human 3 iPS & 3 ES cell lines
in 3 experimental condition at 3 time points}
\format{A data frames with 108 rows and 54 variables:}
\usage{
p.RNAseq6l3c3t
}
\description{
p.RNAseq6l3c3 contains log2(RPKM + 1) values for polyA bulk sequencing of 6 cell lines
in 3 experimental condition at 3 time points.
}
\keyword{datasets}
