% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{pclust-class}
\alias{pclust-class}
\title{pclust}
\description{
parent class of plcustKmeans and plcusltHclust
}
\section{Slots}{

\describe{
\item{\code{pattern}}{pattern found from clusters using cluster2pattern}
}}

