% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pd.ESepiGen4c1l}
\alias{pd.ESepiGen4c1l}
\title{RNAseqing and ChIPSeq of matched genes in differentiated human iPS cells}
\format{
A data frames with 9 rows and 2 variables:
}
\usage{
pd.ESepiGen4c1l
}
\description{
pd.ESepiGen4c1l.4cond contains sample phenotype and experimental information
}
\references{
1. Gifford, C. A. et al. Transcriptional and epigenetic dynamics
during specification of human embryonic stem cells. Cell 153, 1149-1163 (2013).
}
\keyword{datasets}
