% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pdVolcano}
\alias{pdVolcano}
\title{pdVolcano}
\usage{
pdVolcano(
  result,
  FC = 0.2,
  pvalue = NULL,
  subset = NULL,
  filter.inf = FALSE,
  label.num = 5
)
}
\arguments{
\item{result}{result output from projectionDriveR function with PI method selected}

\item{FC}{fold change threshold, default at 0.2}

\item{pvalue}{significance threshold, default set to pvalue stored in projectionDriveR output}

\item{subset}{vector of gene names to subset the plot by}

\item{filter.inf}{remove genes that have pvalues below machine double minimum value}

\item{label.no}{Number of genes to label on either side of the volcano plot, default 5}
}
\value{
A list with weighted and unweighted differential expression metrics
}
\description{
Generate volcano plot and gate genes based on fold change and pvalue, includes vectors that can be used with fast gene set enrichment (fgsea)
}
