% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{create a new project object}
\usage{
project(name, dir = name, path = getwd(), google_proj = name,
  google_path = "projects", repo = name, url = "gitlab.com",
  token = read_token("local/token"), api_version = "v4")
}
\arguments{
\item{name}{project name}

\item{dir}{project directory}

\item{path}{local project location}

\item{google_proj}{google drive location}

\item{google_path}{google drive projects path}

\item{repo}{gitlab repository name}

\item{url}{gitlab server url}

\item{token}{gitlab personal access token}
}
\description{
create a new project object
}
