% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDriveR.R
\name{projectionDriveR}
\alias{projectionDriveR}
\title{projectionDriveR}
\usage{
projectionDriveR(
  cellgroup1,
  cellgroup2,
  loadings,
  loadingsNames = NULL,
  pattern_name,
  alpha,
  display = TRUE,
  normalize_pattern = TRUE
)
}
\arguments{
\item{cellgroup1}{gene x cell count matrix for cell group 1}

\item{cellgroup2}{gene x cell count matrix for cell group 2}

\item{loadings}{A matrix of continuous values defining the features}

\item{loadingsNames}{a vector with names of loading rows. Defaults to rownames.}

\item{pattern_name}{column of loadings for which drivers will be calculated.}

\item{alpha}{confidence value for the bonferroni confidence intervals}

\item{display}{boolean. Whether or not to plot and display confidence intervals}

\item{normalize_pattern}{Boolean. Whether or not to normalize pattern weights.}
}
\value{
A list with weighted mean differences, mean differences, and differential genes that meet the provided signficance threshold.
}
\description{
Calculate the weighted difference in expression between two groups (group1 - group2)
}
