% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionDriveRfun.R
\name{projectionDriveR}
\alias{projectionDriveR}
\title{projectionDriveR}
\usage{
projectionDriveR(
  cellgroup1,
  cellgroup2,
  loadings,
  pattern_name,
  loadingsNames = NULL,
  pvalue = 1e-05,
  display = TRUE,
  normalize_pattern = TRUE,
  mode = "CI"
)
}
\arguments{
\item{cellgroup1}{gene x cell count matrix for cell group 1}

\item{cellgroup2}{gene x cell count matrix for cell group 2}

\item{loadings}{A matrix of continuous values defining the features}

\item{pattern_name}{column of loadings for which drivers will be calculated}

\item{loadingsNames}{a vector with names of loading rows defaults to rownames}

\item{pvalue}{confidence level. Default 1e-5}

\item{display}{boolean. Whether or not to display confidence intervals}

\item{normalize_pattern}{Boolean. Whether or not to normalize pattern weights}

\item{mode}{statistical approach, confidence intervals or pvalues. default CI}
}
\value{
A list with unweighted/weighted mean differences and differential
genes that meet the provided signficance threshold.
}
\description{
Calculate weighted expression difference between two groups (group1 - group2)
}
