% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectR.R
\name{projectR.default}
\alias{projectR.default}
\title{Projection function (default)}
\usage{
\method{projectR}{default}(data = NA, AnnotionObj = NA,
  IDcol = "GeneSymbol", Patterns = NA, NP = NA, full = FALSE,
  model = NA)
}
\arguments{
\item{data}{a dataset to be projected into the pattern space}

\item{AnnotionObj}{an annotion object for data. If NA the rownames of data will be used.}

\item{IDcol}{the column of AnnotionData object corresponding to identifiers matching the type used for GeneWeights}

\item{Patterns}{a matrix of continous values to be projected with unique rownames}

\item{NP}{vector of integers indicating which columns of Patterns object to use. The default of NP = NA will use entire matrix.}

\item{full}{logical indicating whether to return the full model solution. By default only the new pattern object is returned.}

\item{model}{# optional arguements to choose method for projection}
}
\value{
A matrix of sample weights for each input pattern. (if full=TRUE, full model solution is returned)
}
\description{
default version
}
\examples{
   projectR(data=p.ESepiGen4c1l$mRNA.Seq,Patterns=AP.RNAseq6l3c3t,
               AnnotionObj=map.ESepiGen4c1l,IDcol="GeneSymbols")
}
