% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{retinal_patterns}
\alias{retinal_patterns}
\title{CoGAPS patterns learned from the developing mouse retina.}
\format{
A gene (rows) by pattern (column) matrix
}
\usage{
retinal_patterns
}
\description{
CoGAPS patterns learned from the developing mouse retina.
}
\references{
1. Clark, B.S., & Stein-O'Brien G.L., et al. Single-Cell RNA-Seq Analysis of Development Identifies NFI 
Factors as Regulating Mitotic Exit and Late-Born Cell Specification. Cell 102, 1111-1126 (2019).
}
\keyword{datasets}
