% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{rotatoR-class}
\alias{rotatoR-class}
\title{rotatoR}
\description{
class of rotatoR output
}
\section{Slots}{

\describe{
\item{\code{rotatedM}}{rotated matrix that is output of rotatoR function}
}}

