% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateR.R, R/rotatoR.R
\name{rotatoR}
\alias{rotatoR}
\title{rotatoR}
\usage{
rotatoR(x1, y1, x2, y2, basisSET)

rotatoR(x1, y1, x2, y2, basisSET)
}
\arguments{
\item{x1}{a value describing a the coordinate of a point in the first basis. If no values are provided for x2}

\item{y1}{a value describing a the coordinate of a point in the second basis}

\item{x2}{a value describing a the coordinate of the second point in the second basis}

\item{y2}{a value describing a the coordinate of the second point in the second basis}

\item{basisSET}{the basis to be rotated}

\item{x1}{a value describing a the coordinate of a point in the first basis. If no values are provided for x2}

\item{y1}{a value describing a the coordinate of a point in the second basis}

\item{x2}{a value describing a the coordinate of the second point in the second basis}

\item{y2}{a value describing a the coordinate of the second point in the second basis}

\item{basisSET}{the basis to be rotated}
}
\description{
a function for rotating two basis about a point or line in that plain

a function for rotating two basis about a point or line in that plain
}

