% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_me_up.R
\name{set_me_up}
\alias{set_me_up}
\title{Create folder structure}
\usage{
set_me_up(projectname = "Template Project")
}
\arguments{
\item{projectname}{A character string of the project title}
}
\value{
A set of folders created in current directory
}
\description{
This function will create a folder structure that is needed for a data
science project, starting from the current directory.
}
